# fix caffe time command.