--- a/src/caffe/common.cpp
+++ b/src/caffe/common.cpp
@@ -150,6 +150,19 @@ void Caffe::set_random_seed(const unsigned int seed) {
   Get().random_generator_.reset(new RNG(seed));
 }
 
+void Caffe::Synchronize(int device_id) {
+#ifdef USE_GREENTEA
+  DeviceContext& device_context = Caffe::GetDeviceContext(device_id);
+  if ( device_context.backend() == BACKEND_OpenCL ) {
+    viennacl::ocl::context &ctx = viennacl::ocl::get_context(
+                      GetDeviceContext(device_id).id());
+    ctx.get_queue().finish();
+  }
+#else
+  (void) device_id;
+#endif
+}
+
 void Caffe::EnumerateDevices() {
   int cuda_device_count = 0;
   int greentea_device_count = 0;