@naibaf7 2015-06-18
Thanks.
I originally had the finish() on the queue inside the layers. But then I figured out this might hurt asynchronous and multi-GPU execution in the future, so I excluded it.
Your synchronization approach seems sensible and can be triggered when needed only, so I merged it, thanks.